requirements.txt
certifi==2018.1.18
chardet==3.0.4
future==0.16.0
idna==2.6
PyGithub==1.35
PyJWT==1.6.1
python-telegram-bot==7.0.1
requests==2.18.4
urllib3==1.22
